+++ src/main/java/gregtech/integration/jei/utils/AdvancedRecipeWrapper.java
@@ -3,7 +3,10 @@ package gregtech.integration.jei.utils;
 import mezz.jei.api.recipe.IRecipeWrapper;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.audio.PositionedSoundRecord;
+import net.minecraft.client.renderer.GlStateManager;
 import net.minecraft.init.SoundEvents;
+import net.minecraft.item.ItemStack;
+import net.minecraftforge.fml.client.config.GuiUtils;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -22,6 +25,19 @@ public abstract class AdvancedRecipeWrapper implements IRecipeWrapper {
     public void drawInfo(Minecraft minecraft, int recipeWidth, int recipeHeight, int mouseX, int mouseY) {
         for (JeiButton button : buttons) {
             button.render(minecraft, recipeWidth, recipeHeight, mouseX, mouseY);
+            if (button.isHovering(mouseX, mouseY)) {
+                List<String> lines = new ArrayList<>();
+                button.buildTooltip(lines);
+                if (lines.isEmpty())
+                    continue;
+                Minecraft mc = Minecraft.getMinecraft();
+                int width = (int) (mc.displayWidth / 2f + recipeWidth / 2f);
+                int maxWidth = Math.min(200, width - mouseX - 5);
+                GuiUtils.drawHoveringText(ItemStack.EMPTY, lines, mouseX, mouseY,
+                        width,
+                        mc.displayHeight, maxWidth, mc.fontRenderer);
+                GlStateManager.disableLighting();
+            }
         }
     }
